Move marker algebra into vendored packaging
