+++ nab-python/src/nab_python/_vendor/packaging/_markersets.py
@@ -1,37 +1,60 @@
-"""Typed atoms and domain-partition primitives for the marker algebra.
+"""Marker algebra engine: reason about PEP 508 markers as sets of environments.
 
-Parses a marker string (or :class:`Marker`) into a normalised boolean op-tree
-over typed atoms, with the packaging-faithful ``(variable, operator, literal)``
-dispatch, A1 lowering of ``python_version`` onto the ``python_full_version``
-axis, set-valued extras, and opaque ``contains`` atoms. The denotation of a
-value atom is delegated to packaging's own ``_eval_op`` so it matches packaging
-exactly. The decision engine in :mod:`.engine` consumes this tree and reuses the
-per-axis cell partition and per-atom evaluation defined here.
+The private engine behind :class:`~packaging.markersets.MarkerSet`. Parses a
+marker string (or :class:`~packaging.markers.Marker`) into a normalised boolean
+op-tree over typed atoms, with the packaging-faithful
+``(variable, operator, literal)`` dispatch, A1 lowering of ``python_version``
+onto the ``python_full_version`` axis, set-valued extras, and opaque
+``contains`` atoms. The denotation of a value atom is delegated to packaging's
+own ``_eval_op`` so it matches packaging exactly. Decisions run an on-demand
+cell decomposition: every procedure re-partitions the referenced variables'
+domains into cells on which each atom is constant, enumerates the cell product
+under the ``max_cells`` guard, and evaluates the op-tree once per cell.
 """
 
 from __future__ import annotations
 
 import re
 import sys
-from dataclasses import dataclass
-from itertools import pairwise
+from dataclasses import dataclass, replace
+from itertools import pairwise, product
 from typing import TYPE_CHECKING
 
-from .._vendor.packaging._parser import Op, Variable, parse_marker
-from .._vendor.packaging._tokenizer import ParserSyntaxError
-from .._vendor.packaging.markers import (
+from ._parser import Op, Variable, parse_marker
+from ._tokenizer import ParserSyntaxError
+from .markers import (
     InvalidMarker,
     Marker,
     UndefinedComparison,
     UndefinedEnvironmentName,
     _eval_op,
 )
-from .._vendor.packaging.utils import canonicalize_name
-from .._vendor.packaging.version import InvalidVersion, Version
-from .errors import ComplexityLimitError
+from .specifiers import InvalidSpecifier, Specifier
+from .utils import canonicalize_name
+from .version import InvalidVersion, Version
 
 if TYPE_CHECKING:
-    from collections.abc import Iterable, Mapping, Sequence
+    from collections.abc import Iterable, Iterator, Mapping, Sequence
+
+
+class IntractableMarkerSet(ValueError):
+    """The set is too complex to decide within the budget.
+
+    Raised rather than hanging, overflowing the stack, or failing obscurely: an
+    oversized cell product past ``max_cells``, a marker nested past the
+    interpreter stack, or a version literal whose numeric component exceeds the
+    digit parse limit. Subclasses :class:`ValueError` to match packaging's
+    marker exceptions.
+    """
+
+
+class UnserializableMarkerSet(ValueError):
+    """The set has no marker-string spelling.
+
+    Raised, rather than emitting a wrong or masquerading string, for the empty
+    set and for complements whose structure the marker grammar cannot express.
+    Subclasses :class:`ValueError` to match packaging's marker exceptions.
+    """
 
 # Axis kinds. Atoms on the same axis share one cell partition and are each
 # constant on every one of its cells.
@@ -280,22 +303,61 @@ def make_not(node: Formula) -> Formula:
 # ------------------------------------------------------------------- construction
 
 
-def parse(source: str | Marker) -> Formula:
-    """Parse a marker string or :class:`Marker` into the normalised op-tree."""
+def _parse_ast(source: str | Marker) -> list | None:
+    """Dispatch a marker to packaging's parser, or None for an empty marker."""
     if isinstance(source, Marker):
         source = str(source)
     elif not isinstance(source, str):
         msg = f"expected str or Marker, got {type(source).__name__}"
         raise TypeError(msg)
     if not source.strip():
-        return TRUE
+        return None
     try:
-        parsed = parse_marker(source)
+        return parse_marker(source)
     except ParserSyntaxError as exc:
         # A malformed marker raises the public InvalidMarker, as packaging does,
         # not the tokenizer's internal syntax error.
         raise InvalidMarker(str(exc)) from exc
-    return _convert(parsed)
+
+
+def parse(source: str | Marker) -> Formula:
+    """Parse a marker string or :class:`Marker` into the normalised op-tree."""
+    parsed = _parse_ast(source)
+    return TRUE if parsed is None else _convert(parsed)
+
+
+def variable_names(source: str | Marker) -> frozenset[str]:
+    """Return every marker variable ``source`` names, as written.
+
+    Walks the parsed marker collecting the variables its operands name, without
+    building atoms, so a marker the algebra rejects at construction still yields
+    its names. The result over-approximates semantic support.
+    """
+    parsed = _parse_ast(source)
+    if parsed is None:
+        return frozenset()
+
+    names: set[str] = set()
+    _collect_variables(parsed, names)
+    return frozenset(names)
+
+
+def _collect_variables(node: list, names: set[str]) -> None:
+    for item in node:
+        if isinstance(item, str):
+            continue
+        if isinstance(item, list):
+            _collect_variables(item, names)
+        else:
+            lhs, _op, rhs = item
+            if isinstance(lhs, Variable):
+                names.add(lhs.value)
+            if isinstance(rhs, Variable):
+                names.add(rhs.value)
+            elif not isinstance(lhs, Variable) and rhs.value in DOMAIN_REGISTRY:
+                # packaging reads a literal-vs-literal comparison's right operand
+                # as an environment key when it names a variable.
+                names.add(rhs.value)
 
 
 def _convert(node: list) -> Formula:
@@ -411,7 +473,7 @@ def reject_oversized_version_literals(variable: str, literals: Sequence[str]) ->
 
     A numeric component over sys.get_int_max_str_digits() digits makes
     packaging's Version raise a bare ValueError; convert it to the bounded
-    ComplexityLimitError here.
+    IntractableMarkerSet here.
     """
     if is_version_dispatch(variable) and any(
         _oversized_numeric(literal) for literal in literals
@@ -420,7 +482,7 @@ def reject_oversized_version_literals(variable: str, literals: Sequence[str]) ->
             "version literal numeric component exceeds the "
             f"{sys.get_int_max_str_digits()}-digit parse limit"
         )
-        raise ComplexityLimitError(msg)
+        raise IntractableMarkerSet(msg)
 
 
 def _reject_mint_overflow(literals: Sequence[str]) -> None:
@@ -440,7 +502,7 @@ def _reject_mint_overflow(literals: Sequence[str]) -> None:
             "version literal numeric component leaves no headroom under the "
             f"{limit}-digit parse limit"
         )
-        raise ComplexityLimitError(msg)
+        raise IntractableMarkerSet(msg)
 
 
 def _reject_undefined_operator(
@@ -614,7 +676,7 @@ def _elevate_epochs(
             elevated.append(f"{epoch}!{text}")
             if len(elevated) > max_cells:
                 msg = f"version pool exceeds max_cells={max_cells}"
-                raise ComplexityLimitError(msg)
+                raise IntractableMarkerSet(msg)
     return elevated
 
 
@@ -651,7 +713,7 @@ def _dedupe_candidates(
         ordered.append(candidate)
         if len(ordered) > max_cells:
             msg = f"value candidate set exceeds max_cells={max_cells}"
-            raise ComplexityLimitError(msg)
+            raise IntractableMarkerSet(msg)
     return ordered
 
 
@@ -696,7 +758,7 @@ def _value_candidates(
     reject_oversized_version_literals(variable, literals)
     if _reduce_work_exceeds(variable, literals, len(atoms), max_cells):
         msg = f"axis work over {len(atoms)} atoms exceeds max_cells={max_cells}"
-        raise ComplexityLimitError(msg)
+        raise IntractableMarkerSet(msg)
 
     candidates: list[str] = []
     raw_kind = DOMAIN_REGISTRY[variable]
@@ -715,7 +777,7 @@ def _value_candidates(
             spent += _substring_cost(atom.literal)
             if spent > max_cells:
                 msg = f"substring enumeration exceeds max_cells={max_cells}"
-                raise ComplexityLimitError(msg)
+                raise IntractableMarkerSet(msg)
             candidates.extend(_membership_candidates(atom))
 
     if is_version_dispatch(variable):
@@ -749,7 +811,7 @@ def _reduce_cells(
         msg = (
             f"axis work {len(points)}x{len(atoms)} atoms exceeds max_cells={max_cells}"
         )
-        raise ComplexityLimitError(msg)
+        raise IntractableMarkerSet(msg)
 
     representatives: dict[tuple, object] = {}
     for point in points:
@@ -787,7 +849,7 @@ def partition_set_axis(atoms: Sequence[Atom], max_cells: int) -> list[Cell]:
     count = len(names)
     if (1 << count) > max_cells:
         msg = f"set powerset over {count} names exceeds max_cells={max_cells}"
-        raise ComplexityLimitError(msg)
+        raise IntractableMarkerSet(msg)
     subsets = [
         frozenset(names[i] for i in range(count) if mask & (1 << i))
         for mask in range(1 << count)
@@ -817,7 +879,7 @@ def guarded_product_size(sizes: Iterable[int], max_cells: int) -> int:
         total *= size
         if total > max_cells:
             msg = f"cell product exceeds max_cells={max_cells}"
-            raise ComplexityLimitError(msg)
+            raise IntractableMarkerSet(msg)
     return total
 
 
@@ -858,3 +920,349 @@ def evaluate_atom(atom: Atom, env: Mapping[str, object]) -> bool:
     if atom.kind == AXIS_SET:
         return atom.holds(as_name_set(_require(env, atom.origin)))
     return atom.holds(atom.literal in _require(env, atom.variable))  # type: ignore[operator]
+
+
+_MISSING = object()
+
+
+# --------------------------------------------------------------------- walking
+
+
+def _walk(node: Formula, out: list[Atom]) -> None:
+    if isinstance(node, AtomLeaf):
+        out.append(node.atom)
+    elif isinstance(node, NotNode):
+        _walk(node.child, out)
+    elif isinstance(node, (AndNode, OrNode)):
+        for child in node.children:
+            _walk(child, out)
+
+
+def collect_atoms(node: Formula) -> list[Atom]:
+    """Every atom mentioned by a tree, in encounter order."""
+    out: list[Atom] = []
+    _walk(node, out)
+    return out
+
+
+def reject_oversized_literals(node: Formula, env: Mapping[str, object]) -> None:
+    """Raise the bounded guard for an oversized version literal or env value.
+
+    Runs before either could reach packaging's Version and raise a bare
+    ValueError.
+    """
+    for atom in collect_atoms(node):
+        value = _atom_env_value(atom, env)
+        if value is not _MISSING:
+            reject_oversized_version_literals(atom.variable, (atom.literal, str(value)))
+
+
+def variables_of(node: Formula) -> frozenset[str]:
+    """Return the variables a tree references, as written."""
+    return frozenset(atom.origin for atom in collect_atoms(node))
+
+
+def membership_literals_of(node: Formula) -> frozenset[tuple[str, str]]:
+    """Return the ``(variable, canonical name)`` set-memberships a tree tests."""
+    return frozenset(
+        (atom.origin, atom.literal)
+        for atom in collect_atoms(node)
+        if atom.kind == AXIS_SET
+    )
+
+
+def unprovided_variables(node: Formula, env: Mapping[str, object]) -> set[str]:
+    """Return the referenced variables an environment supplies no value for."""
+    return {
+        atom.origin
+        for atom in collect_atoms(node)
+        if _atom_env_value(atom, env) is _MISSING
+    }
+
+
+def _atoms_by_axis(atoms: list[Atom]) -> dict[tuple[str, ...], list[Atom]]:
+    grouped: dict[tuple[str, ...], list[Atom]] = {}
+    seen: dict[tuple[str, ...], set[Atom]] = {}
+    for atom in atoms:
+        axis = atom.axis()
+        known = seen.setdefault(axis, set())
+        if atom not in known:
+            known.add(atom)
+            grouped.setdefault(axis, []).append(atom)
+    return grouped
+
+
+# ------------------------------------------------------------------- decisions
+
+
+def _eval_cell(node: Formula, truth: Mapping[Atom, bool]) -> bool:
+    if isinstance(node, BoolConst):
+        return node.value
+    if isinstance(node, AtomLeaf):
+        return truth[node.atom]
+    if isinstance(node, NotNode):
+        return not _eval_cell(node.child, truth)
+    if isinstance(node, AndNode):
+        return all(_eval_cell(child, truth) for child in node.children)
+    return any(_eval_cell(child, truth) for child in node.children)
+
+
+def _satisfying_cells(
+    node: Formula, max_cells: int
+) -> Iterator[dict[tuple[str, ...], Cell]]:
+    atoms = collect_atoms(node)
+    grouped = _atoms_by_axis(atoms)
+    if not grouped:
+        if _eval_cell(node, {}):
+            yield {}
+        return
+
+    axes = list(grouped)
+    atomlists = [grouped[axis] for axis in axes]
+    partitions = [
+        partition_axis(axis, atoms, max_cells)
+        for axis, atoms in zip(axes, atomlists, strict=True)
+    ]
+
+    # The enumeration walks the whole op-tree once per cell, so guard the cell
+    # product times the leaf-occurrence count: a marker that repeats atoms inflates
+    # the walk without inflating the distinct-atom count or the cell product.
+    leaf_occurrences = len(atoms)
+    guarded_product_size(
+        (*(len(part) for part in partitions), leaf_occurrences), max_cells
+    )
+
+    for combo in product(*partitions):
+        truth: dict[Atom, bool] = {
+            atom: value
+            for atoms, cell in zip(atomlists, combo, strict=True)
+            for atom, value in zip(atoms, cell.vector, strict=True)
+        }
+        if _eval_cell(node, truth):
+            yield dict(zip(axes, combo, strict=True))
+
+
+def is_empty(node: Formula, max_cells: int) -> bool:
+    """Whether a tree denotes the empty set."""
+    return next(_satisfying_cells(node, max_cells), _MISSING) is _MISSING
+
+
+def witness(node: Formula, max_cells: int) -> dict[str, str | frozenset[str]] | None:
+    """Return a concrete environment satisfying a tree, or ``None`` if none is found.
+
+    The returned environment is verified against the tree before it is returned.
+    ``None`` is returned for the empty set. The search over ``contains`` atoms
+    is incomplete, so ``None`` may also be returned for a non-empty set when the
+    concrete-string constraints on one variable (a value atom, one or more
+    ``contains`` atoms, or a mix) have no jointly realisable cell representative.
+    ``python_version`` and ``python_full_version`` share one axis, so those
+    constraints can sit on different variables.
+    """
+    for cell in _satisfying_cells(node, max_cells):
+        env = _materialize(cell)
+        if evaluate_tree(node, env):
+            return env
+    return None
+
+
+def _materialize(
+    cell: Mapping[tuple[str, ...], Cell],
+) -> dict[str, str | frozenset[str]]:
+    env: dict[str, str | frozenset[str]] = {}
+    contains: dict[str, list[tuple[str, bool]]] = {}
+    for axis, piece in cell.items():
+        kind = axis[0]
+        if kind == AXIS_VALUE:
+            env[axis[1]] = str(piece.point)
+        elif kind == AXIS_SET:
+            env[axis[1]] = frozenset(piece.point)  # type: ignore[arg-type]
+        else:
+            contains.setdefault(axis[1], []).append((axis[2], bool(piece.point)))
+
+    for variable, items in contains.items():
+        if variable in env:
+            continue
+        if variable == "python_version" and "python_full_version" in env:
+            env[variable] = derive_major_minor(str(env["python_full_version"]))
+            continue
+        env[variable] = "".join(sorted(lit for lit, present in items if present))
+    return env
+
+
+# --------------------------------------------------------------------- restrict
+
+
+def _atom_env_value(atom: Atom, env: Mapping[str, object]) -> object:
+    """Return the env value the atom reads, or ``_MISSING`` when unprovided."""
+    if atom.derive_mm:
+        for key in ("python_full_version", "python_version"):
+            if key in env:
+                return env[key]
+        return _MISSING
+    key = atom.origin if atom.kind == AXIS_SET else atom.variable
+    return env.get(key, _MISSING)
+
+
+def _restrict_value(atom: Atom, env: Mapping[str, object]) -> bool | None:
+    """Return the atom's truth under ``env``, or ``None`` when unprovided."""
+    value = _atom_env_value(atom, env)
+    if value is _MISSING:
+        return None
+    if atom.kind == AXIS_SET:
+        return atom.holds(as_name_set(value))
+    if atom.kind == AXIS_CONTAINS:
+        return atom.holds(atom.literal in value)  # type: ignore[operator]
+    return atom.holds(value)
+
+
+def _restrict_atom(leaf: AtomLeaf, env: Mapping[str, object]) -> Formula:
+    resolved = _restrict_value(leaf.atom, env)
+    if resolved is None:
+        return leaf
+    return TRUE if resolved else FALSE
+
+
+def restrict_tree(node: Formula, env: Mapping[str, object]) -> Formula:
+    """Substitute the provided variables, leaving the rest as a residual."""
+    if isinstance(node, BoolConst):
+        return node
+    if isinstance(node, AtomLeaf):
+        return _restrict_atom(node, env)
+    if isinstance(node, NotNode):
+        return make_not(restrict_tree(node.child, env))
+    if isinstance(node, AndNode):
+        return make_and(restrict_tree(child, env) for child in node.children)
+    return make_or(restrict_tree(child, env) for child in node.children)
+
+
+# ------------------------------------------------------------------- evaluate
+
+
+def evaluate_tree(node: Formula, env: Mapping[str, object]) -> bool:
+    """Evaluate a tree against a full environment (extras are sets)."""
+    if isinstance(node, BoolConst):
+        return node.value
+    if isinstance(node, AtomLeaf):
+        return evaluate_atom(node.atom, env)
+    if isinstance(node, NotNode):
+        return not evaluate_tree(node.child, env)
+    if isinstance(node, AndNode):
+        return all(evaluate_tree(child, env) for child in node.children)
+    return any(evaluate_tree(child, env) for child in node.children)
+
+
+# ------------------------------------------------------------------ serialise
+
+
+def _builds_specifier(op: str, literal: str) -> bool:
+    try:
+        Specifier(f"{op}{literal}")
+    except InvalidSpecifier:
+        return False
+    return True
+
+
+def _complement_version(atom: Atom, op: str, var: str) -> Formula:
+    # Excluded middle holds only for ==/!= on a pure-version axis; ordered
+    # comparisons have the prerelease hole, and the twins can hold a non-version,
+    # so neither complements to a single atom.
+    if op in ("==", "!=") and is_pure_version(var) and not atom.swapped:
+        return AtomLeaf(replace(atom, op="!=" if op == "==" else "=="))
+    msg = f"cannot complement version atom on {var!r}"
+    raise UnserializableMarkerSet(msg)
+
+
+def _complement_string(atom: Atom, op: str) -> Formula:
+    if op in ("==", ">=", "<="):
+        return AtomLeaf(replace(atom, op="!="))
+    if op == "!=":
+        return AtomLeaf(replace(atom, op="=="))
+    if op == "in":
+        return AtomLeaf(replace(atom, op="not in"))
+    if op == "not in":
+        return AtomLeaf(replace(atom, op="in"))
+    # < and > are constant-false on a string variable, so the complement is all.
+    return TRUE
+
+
+def _complement_leaf(atom: Atom) -> Formula:
+    if atom.kind in (AXIS_SET, AXIS_CONTAINS):
+        return AtomLeaf(replace(atom, positive=not atom.positive))
+    op, var = atom.op, atom.variable
+    if is_version_dispatch(var) and _builds_specifier(op, atom.literal):
+        return _complement_version(atom, op, var)
+    return _complement_string(atom, op)
+
+
+def to_nnf(node: Formula) -> Formula:
+    """Push complements down to the leaves (negation normal form)."""
+    if isinstance(node, AtomLeaf):
+        return node
+    if isinstance(node, AndNode):
+        return make_and(to_nnf(child) for child in node.children)
+    if isinstance(node, OrNode):
+        return make_or(to_nnf(child) for child in node.children)
+    if isinstance(node, NotNode):
+        return _negate(node.child)
+    msg = "a bare constant cannot reach to_nnf"  # pragma: no cover
+    raise RuntimeError(msg)  # pragma: no cover
+
+
+def _negate(node: Formula) -> Formula:
+    if isinstance(node, AtomLeaf):
+        return _complement_leaf(node.atom)
+    if isinstance(node, AndNode):
+        return make_or(_negate(child) for child in node.children)
+    if isinstance(node, OrNode):
+        return make_and(_negate(child) for child in node.children)
+    if isinstance(node, NotNode):
+        return to_nnf(node.child)
+    msg = "a bare constant cannot reach _negate"  # pragma: no cover
+    raise RuntimeError(msg)  # pragma: no cover
+
+
+def _quote(literal: str) -> str:
+    """Spell a literal as a marker string, picking the quote the grammar allows.
+
+    A PEP 508 literal is delimited by one quote style and cannot contain that
+    style, so a literal carrying a double-quote is spelled with single quotes.
+    """
+    if '"' not in literal:
+        return f'"{literal}"'
+    if "'" not in literal:
+        return f"'{literal}'"
+    # A literal carrying both quote styles has no marker spelling. Marker
+    # literals only ever arrive through the exclusive-quote grammar, so a value
+    # holding both is unreachable from any parsed input.
+    msg = f"literal {literal!r} has no marker-string quoting"  # pragma: no cover
+    raise UnserializableMarkerSet(msg)  # pragma: no cover
+
+
+def _render_atom(atom: Atom) -> str:
+    if atom.kind == AXIS_SET and atom.origin == "extra":
+        op = "==" if atom.positive else "!="
+        return f"extra {op} {_quote(atom.literal)}"
+    if atom.kind in (AXIS_SET, AXIS_CONTAINS):
+        op = "in" if atom.positive else "not in"
+        return f"{_quote(atom.literal)} {op} {atom.origin}"
+    if atom.swapped:
+        return f"{_quote(atom.literal)} {atom.op} {atom.origin}"
+    return f"{atom.origin} {atom.op} {_quote(atom.literal)}"
+
+
+def _paren(node: Formula) -> str:
+    if isinstance(node, AtomLeaf):
+        return _render_atom(node.atom)
+    return f"({serialize(node)})"
+
+
+def serialize(node: Formula) -> str:
+    """Render a negation-normal-form tree to a marker string."""
+    if isinstance(node, AtomLeaf):
+        return _render_atom(node.atom)
+    if isinstance(node, AndNode):
+        return " and ".join(_paren(child) for child in node.children)
+    if isinstance(node, OrNode):
+        return " or ".join(_paren(child) for child in node.children)
+    msg = "a bare constant has no marker-atom spelling"  # pragma: no cover
+    raise RuntimeError(msg)  # pragma: no cover
