+++ nab-python/src/nab_python/_vendor/packaging/markersets.py
@@ -1,21 +1,47 @@
-"""The public :class:`MarkerSet`: a marker's denotation as a set of environments."""
+"""The public :class:`MarkerSet`: a marker's denotation as a set of environments.
+
+A :class:`MarkerSet` is the denotation of a PEP 508 marker as a set of
+environments, the marker-side counterpart of :class:`~packaging.ranges.VersionRange`.
+It holds the states a marker string cannot (the universal set of an absent
+marker, the empty set of a contradiction, and complements the grammar cannot
+spell), and reconciles with the grammar at exactly one boundary,
+:meth:`~MarkerSet.to_marker_string`, which may return ``None`` or raise. The
+engine is on-demand cell decomposition, guarded by ``max_cells``.
+"""
 
 from __future__ import annotations
 
 from functools import wraps
 from typing import TYPE_CHECKING, ParamSpec, TypeVar
 
-from . import atoms, engine
-from .errors import ComplexityLimitError, UnserializableSetError
+from . import _markersets
+from ._markersets import (
+    IntractableMarkerSet,
+    UnserializableMarkerSet,
+    variable_names,
+)
 
 if TYPE_CHECKING:
     from collections.abc import Callable, Mapping
     from collections.abc import Set as AbstractSet
+    from typing import Literal
 
-    from .._vendor.packaging.markers import Marker
-    from .atoms import Formula
+    from ._markersets import Formula
+    from .markers import Marker
 
 DEFAULT_MAX_CELLS = 100_000
+
+__all__ = [
+    "IntractableMarkerSet",
+    "MarkerSet",
+    "UnserializableMarkerSet",
+    "variable_names",
+]
+
+
+def __dir__() -> list[str]:
+    return __all__
+
 
 _P = ParamSpec("_P")
 _R = TypeVar("_R")
@@ -26,7 +52,7 @@ def _bounded(method: Callable[_P, _R]) -> Callable[_P, _R]:
 
     A tree walk recurses as deep as the marker nests, so a marker nested past the
     interpreter's stack raises :class:`RecursionError`. The public methods it
-    decorates report it as :class:`ComplexityLimitError`, the one bounded
+    decorates report it as :class:`IntractableMarkerSet`, the one bounded
     failure the algebra promises on pathological input.
     """
 
@@ -36,7 +62,7 @@ def _bounded(method: Callable[_P, _R]) -> Callable[_P, _R]:
             return method(*args, **kwargs)
         except RecursionError as exc:
             msg = "marker nests too deeply to decide"
-            raise ComplexityLimitError(msg) from exc
+            raise IntractableMarkerSet(msg) from exc
 
     return wrapper
 
@@ -66,67 +92,73 @@ class MarkerSet:
     ) -> MarkerSet:
         """Build the set of a marker string or :class:`Marker`."""
         _check_max_cells(max_cells)
-        return cls(atoms.parse(marker), max_cells)
+        return cls(_markersets.parse(marker), max_cells)
 
     @classmethod
     def true(cls, *, max_cells: int = DEFAULT_MAX_CELLS) -> MarkerSet:
         """Return the universal set: every environment (an absent marker)."""
         _check_max_cells(max_cells)
-        return cls(atoms.TRUE, max_cells)
+        return cls(_markersets.TRUE, max_cells)
 
     @classmethod
     def false(cls, *, max_cells: int = DEFAULT_MAX_CELLS) -> MarkerSet:
         """Return the empty set: no environment."""
         _check_max_cells(max_cells)
-        return cls(atoms.FALSE, max_cells)
+        return cls(_markersets.FALSE, max_cells)
 
     # ---- algebra
 
     def __and__(self, other: MarkerSet) -> MarkerSet:
-        return MarkerSet(atoms.make_and((self._tree, other._tree)), self._max_cells)
+        return MarkerSet(
+            _markersets.make_and((self._tree, other._tree)), self._max_cells
+        )
 
     def __or__(self, other: MarkerSet) -> MarkerSet:
-        return MarkerSet(atoms.make_or((self._tree, other._tree)), self._max_cells)
+        return MarkerSet(
+            _markersets.make_or((self._tree, other._tree)), self._max_cells
+        )
 
     def complement(self) -> MarkerSet:
         """Return the set of environments this set excludes."""
-        return MarkerSet(atoms.make_not(self._tree), self._max_cells)
+        return MarkerSet(_markersets.make_not(self._tree), self._max_cells)
 
     # ---- decision procedures
 
     @_bounded
     def is_empty(self) -> bool:
         """Whether the set is unsatisfiable."""
-        return engine.is_empty(self._tree, self._max_cells)
+        return _markersets.is_empty(self._tree, self._max_cells)
 
     @_bounded
     def is_tautology(self) -> bool:
         """Whether the set is every environment."""
-        return engine.is_empty(atoms.make_not(self._tree), self._max_cells)
+        return _markersets.is_empty(
+            _markersets.make_not(self._tree), self._max_cells
+        )
 
     @_bounded
     def is_disjoint(self, other: MarkerSet) -> bool:
         """Whether the two sets share no environment."""
-        return engine.is_empty(
-            atoms.make_and((self._tree, other._tree)), self._max_cells
+        return _markersets.is_empty(
+            _markersets.make_and((self._tree, other._tree)), self._max_cells
         )
 
     @_bounded
     def implies(self, other: MarkerSet) -> bool:
         """Whether every environment in this set is in ``other``."""
-        return engine.is_empty(
-            atoms.make_and((self._tree, atoms.make_not(other._tree))),
+        return _markersets.is_empty(
+            _markersets.make_and((self._tree, _markersets.make_not(other._tree))),
             self._max_cells,
         )
 
     @_bounded
     def equivalent(self, other: MarkerSet) -> bool:
         """Whether the two sets denote the same environments."""
-        return engine.is_empty(
-            atoms.make_and((self._tree, atoms.make_not(other._tree))),
+        return _markersets.is_empty(
+            _markersets.make_and((self._tree, _markersets.make_not(other._tree))),
             self._max_cells,
-        ) and engine.is_empty(
-            atoms.make_and((other._tree, atoms.make_not(self._tree))),
+        ) and _markersets.is_empty(
+            _markersets.make_and((other._tree, _markersets.make_not(self._tree))),
             self._max_cells,
         )
 
@@ -137,7 +169,7 @@ class MarkerSet:
         self,
         env: Mapping[str, str | AbstractSet[str]],
         *,
-        on_unknown_variable: str = "residual",
+        on_unknown_variable: Literal["residual", "error"] = "residual",
     ) -> MarkerSet:
         """Substitute the provided variables, returning a residual set.
 
@@ -152,30 +184,27 @@ class MarkerSet:
             )
             raise ValueError(msg)
         if on_unknown_variable == "error":
-            missing = engine.unprovided_variables(self._tree, env)
+            missing = _markersets.unprovided_variables(self._tree, env)
             if missing:
                 msg = f"restrict() has no value for {sorted(missing)}"
                 raise ValueError(msg)
-        engine.reject_oversized_literals(self._tree, env)
-        return MarkerSet(engine.restrict_tree(self._tree, env), self._max_cells)
-
-    @_bounded
-    def variables(self) -> frozenset[str]:
-        """Return the variables the set references, as written."""
-        return engine.variables_of(self._tree)
+        _markersets.reject_oversized_literals(self._tree, env)
+        return MarkerSet(
+            _markersets.restrict_tree(self._tree, env), self._max_cells
+        )
 
     @_bounded
     def membership_literals(self) -> frozenset[tuple[str, str]]:
         """Return the ``(variable, canonical name)`` set-memberships the set tests."""
-        return engine.membership_literals_of(self._tree)
+        return _markersets.membership_literals_of(self._tree)
 
     # ---- evaluation and witness
 
     @_bounded
     def evaluate(self, env: Mapping[str, str | AbstractSet[str]]) -> bool:
         """Whether a full environment is in the set (extras are sets)."""
-        engine.reject_oversized_literals(self._tree, env)
-        return engine.evaluate_tree(self._tree, env)
+        _markersets.reject_oversized_literals(self._tree, env)
+        return _markersets.evaluate_tree(self._tree, env)
 
     @_bounded
     def witness(self) -> dict[str, str | frozenset[str]] | None:
@@ -188,7 +217,7 @@ class MarkerSet:
         cell representative. ``python_version`` and ``python_full_version``
         share one axis, so those constraints can sit on different variables.
         """
-        return engine.witness(self._tree, self._max_cells)
+        return _markersets.witness(self._tree, self._max_cells)
 
     # ---- serialisation
 
@@ -198,22 +227,22 @@ class MarkerSet:
 
         ``None`` means the universal set (no marker). The empty set, and any set
         whose complement structure the marker grammar cannot express, raise
-        :class:`UnserializableSetError` rather than emit a wrong string. The produced
-        string is verified equivalent to this set before it is returned.
+        :class:`UnserializableMarkerSet` rather than emit a wrong string. The
+        produced string is verified equivalent to this set before it is returned.
         """
         if self.is_tautology():
             return None
         if self.is_empty():
             msg = "the empty set has no marker string"
-            raise UnserializableSetError(msg)
+            raise UnserializableMarkerSet(msg)
 
-        text = engine.serialize(engine.to_nnf(self._tree))
+        text = _markersets.serialize(_markersets.to_nnf(self._tree))
         rebuilt = MarkerSet.from_marker(text, max_cells=self._max_cells)
         if not self.equivalent(rebuilt):  # pragma: no cover
             # A last-resort guard: the per-atom complements are sound by
             # construction, so a non-equivalent round-trip is unreachable.
             msg = "serialisation is not round-trip sound"
-            raise UnserializableSetError(msg)
+            raise UnserializableMarkerSet(msg)
         return text
 
     @_bounded
